pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "Over Engineered Todo List"
readme = "README.md"
authors = ["Robert"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
pydantic = "^2.3"
alembic = "^1.12.0"
tenacity = "^8.2.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.21"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.0.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.3.1"
httpx = "^0.24.1"
pytest = "^7.4.0"
ruff = "^0.0.287"
python-dotenv = "^1.0.0"
